+++ domain/src/main/java/com/example/domain/contracts/IRepository.kt
@@ -4,6 +4,6 @@ import com.example.domain.entities.Movie
 import kotlinx.coroutines.flow.Flow
 
 interface IRepository<T> {
-    suspend fun getById(id: Int): Flow<T>
+    suspend fun getById(id: Long): Flow<T>
     suspend fun search(query: String): Flow<List<T>>
 }
